requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
alembic==1.13.3
jinja2==3.1.4
python-multipart==0.0.12
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
httpx==0.27.2
apscheduler==3.10.4
itsdangerous==2.2.0
starlette==0.38.4
aiosqlite==0.20.0
